findWorkFlowRequestsByApproverId filtered based on date range and actiontype with approve or reject and showing the last step, incase the approver is same for multiple steps. (#153)
* Implementing approval matrix and listing WF reqeusts based on the approver

* Implementing approval matrix and listing WF reqeusts based on the approver

* implementing getallwf requests for requestor

* implementing getallwf requests for requestor

* addressed single process step having multiple approvers

* including complete details for the approvers wf requests

* aligining the update and add request of TimeSheet

* including TSInfo and Timeentry details in the WFRequest response

* moving processWorkfflow to WorkFlowService

* deducting total rejected time from the total submitted time

* Fixing null as string for the approvedtime

* Modified the getworkflows details based on the approver id and changed the reqeust structure for the processworkFlow

Modified the getworkflows details based on the approver id and changed the reqeust structure for the processworkFlow

* disabling the submitTS in live mode

* pushing get wf request history details based on the tsid

* fixing duplicate data issue

* implementing WF changes and editprojectbytemplate

* incorporated modified wf changes

* wfr;s filtering by current step

wfr;s filtering by current step

* refactored the wf history for the ts

refactored the wf history for the ts

* adding new api to provide all active and inactive wfr's based on the approver ids

adding new api to provide all active and inactive wfr's based on the approver ids

* removing unused import

* Implemented new design changes, w.r.t wf and its comments.

Implemented new design changes, w.r.t wf and its comments.

* turning off test mode

* adding multiple screens to accesstype and deletion based on accesstype and clientid

adding multiple screens to accesstype and deletion based on accesstype and clientid.
fixing the getallprocessteps issue in multiple approvers case for a single step.

* modifying the joiner for the multiple approvers

* modifying the joiner for the multiple approvers

modifying the joiner for the multiple approvers

* included superadmin attribute in the authenticate api

included superadmin attribute in the authenticate api and checking the client status along with the user status, returning active flag as false if any of the status is 0

* Removed the setting of status id in the TSTimeentries as discussed with Bharath

Removed the setting of  status id in the TSTimeentries as discussed with Bharath

* fixing logging issue

* fixing logging issue

* Marking the current step as true for the final step(Rejecte. Revised, Approved(last step)).

Marking the current step as true inthe final Action taken step(final action can Rejected. Revised, Approved  in last step).

* Marking the current step as true inthe final Action taken step(final action can Rejected. Revised, Approved  in last step).

Marking the current step as true inthe final Action taken step(final action can Rejected. Revised, Approved  in last step).

* persisting the ts status manually

* fixing the persistence issue and  process step by client and projectid filter

* implemented reports and refactored add accesstype return type

* disabling the test mode

* findWorkFlowRequestsByApproverId filtered based on date range and actiontype

findWorkFlowRequestsByApproverId filtered based on date range and actiontype with approve or reject and showing the last step, incase the approver is same for multiple steps.

Co-authored-by: ranganathprn <ranganathprn@gmail.com>